+++ matminer/featurizers/site.py
@@ -82,1357 +82,7 @@ with open(os.path.join(os.path.dirname(__file__),
 
 
 
-class IntersticeDistribution(BaseFeaturizer):
-    """
-    Interstice distribution in the neighboring cluster around an atom site.
 
-    The interstices are categorized to distance, area and volume interstices.
-    Each of these metrics is a measures of the relative amount of empty space
-    around each atom as determined using atomic sphere models. The distance
-    interstice is the fraction of a bonding line unoccupied by the atom spheres;
-    The area interstice is the unoccupied area within the triangulated surface
-    formed by atom triplets in convex hull formed by neighbors, and the volume
-    interstice is the unoccupied portion of a tetrahedron formed between the
-    central atom and neighbor atom triplets. Please refer to the original paper
-    for more details (Wang et al. Nat Commun 10, 5537 (2019))
-
-    For amorphous alloys (metallic glasses), the coordination environments are
-    anisotropic, which can be reflected in the inequality of the interstices
-    present around an atom. To describe the anisotropy, here we derive statistics
-    of the interstices to featurize the interstice distribution around the atom.
-    Other methods can be grouping the interstices into histogram grids of fixed
-    bins and the features are then a vector of the values of the histograms.
-
-    User note:
-    This class is particularly designed for featuring the site-specific packing
-    heterogeneity in metallic glasses, especially the all-metallic-element ones.
-    If non-metallic-elements are present in the structures, the interstice
-    estimates may have larger deviation from actual values (despite this
-    deviation is systematic and thus the interstice estimates can still be
-    used to represent the packing heterogeneity).
-
-    Args:
-        cutoff (float): cutoff distance in determining the potential
-            neighbors for Voronoi tessellation analysis. (default: 6.5)
-        interstice_types (str or [str]): interstice distribution types,
-            support sub-list of ['dist', 'area', 'vol'].
-        stats ([str]): statistics of distance/area/volume interstices.
-        radius_type (str): source of radius estimate. (default: "MiracleRadius")
-    """
-
-    def __init__(self, cutoff=6.5, interstice_types=None, stats=None, radius_type="MiracleRadius"):
-        self.cutoff = cutoff
-        self.interstice_types = ["dist", "area", "vol"] if interstice_types is None else interstice_types
-        if isinstance(self.interstice_types, str):
-            self.interstice_types = [self.interstice_types]
-        if all(t not in self.interstice_types for t in ["dist", "area", "vol"]):
-            raise ValueError("interstice_types only support sub-list of " "['dist', 'area', 'vol']")
-        self.stats = ["mean", "std_dev", "minimum", "maximum"] if stats is None else stats
-        self.radius_type = radius_type
-
-    def featurize(self, struct, idx):
-        """
-        Get interstice distribution fingerprints of site with given index in
-        input structure.
-        Args:
-            struct (Structure): Pymatgen Structure object.
-            idx (int): index of target site in structure.
-        Returns:
-            interstice_fps ([float]): Interstice distribution fingerprints.
-        """
-        interstice_fps = list()
-
-        # Get the nearest neighbors using Voronoi tessellation
-        n_w = VoronoiNN(cutoff=self.cutoff).get_voronoi_polyhedra(struct, idx).values()
-
-        nn_coords = np.array([nn["site"].coords for nn in n_w])
-
-        # Get center atom's radius and its nearest neighbors' radii
-        center_r = MagpieData().get_elemental_properties([struct[idx].specie], self.radius_type)[0] / 100
-        nn_els = [nn["site"].specie for nn in n_w]
-        nn_rs = np.array(MagpieData().get_elemental_properties(nn_els, self.radius_type)) / 100
-
-        # Get indices of atoms forming the simplices of convex hull
-        convex_hull_simplices = ConvexHull(nn_coords).simplices
-
-        if "dist" in self.interstice_types:
-            nn_dists = [nn["face_dist"] * 2 for nn in n_w]
-            interstice_dist_list = IntersticeDistribution.analyze_dist_interstices(center_r, nn_rs, nn_dists)
-            interstice_fps += [PropertyStats().calc_stat(interstice_dist_list, stat) for stat in self.stats]
-
-        if "area" in self.interstice_types:
-            interstice_area_list = IntersticeDistribution.analyze_area_interstice(
-                nn_coords, nn_rs, convex_hull_simplices
-            )
-            interstice_fps += [PropertyStats().calc_stat(interstice_area_list, stat) for stat in self.stats]
-
-        if "vol" in self.interstice_types:
-            interstice_vol_list = IntersticeDistribution.analyze_vol_interstice(
-                struct[idx].coords, nn_coords, center_r, nn_rs, convex_hull_simplices
-            )
-            interstice_fps += [PropertyStats().calc_stat(interstice_vol_list, stat) for stat in self.stats]
-        return interstice_fps
-
-    @staticmethod
-    def analyze_dist_interstices(center_r, nn_rs, nn_dists):
-        """Analyze the distance interstices between center atom and neighbors.
-        Args:
-            center_r (float): central atom's radius.
-            nn_rs ([float]): Nearest Neighbors' radii.
-            nn_dists ([float]): Nearest Neighbors' distances.
-        Returns:
-            dist_interstice_list ([float]): Distance interstice list.
-        """
-        dist_interstice_list = list()
-        for nn_dist, nn_r in zip(nn_dists, nn_rs):
-            dist_interstice_list.append(nn_dist / (center_r + nn_r) - 1)
-        return dist_interstice_list
-
-    @staticmethod
-    def analyze_area_interstice(nn_coords, nn_rs, convex_hull_simplices):
-        """Analyze the area interstices in the neighbor convex hull facets.
-        Args:
-            nn_coords (array-like, shape (N, 3)): Nearest Neighbors' coordinates
-            nn_rs ([float]): Nearest Neighbors' radii.
-            convex_hull_simplices (array-like, shape (M, 3)): Indices of points
-                forming the simplicial facets of convex hull.
-        Returns:
-            area_interstice_list ([float]): Area interstice list.
-        """
-        area_interstice_list = list()
-
-        triplet_set = [(0, 1, 2), (1, 0, 2), (2, 0, 1)]
-        for facet_indices in convex_hull_simplices:
-            facet_coords = nn_coords[facet_indices]
-            facet_rs = nn_rs[facet_indices]
-            triangle_angles = list()
-            for triplet in triplet_set:
-                a = facet_coords[triplet[1]] - facet_coords[triplet[0]]
-                b = facet_coords[triplet[2]] - facet_coords[triplet[0]]
-                t_a = np.arccos(np.dot(a, b) / (np.linalg.norm(a) * np.linalg.norm(b)))
-                triangle_angles.append(t_a)
-
-            # calculate neighbors' packed area in the facet
-            packed_area = 0
-            for t_a, nn_r in zip(triangle_angles, facet_rs):
-                packed_area += t_a / 2 * pow(nn_r, 2)
-
-            triangle_area = 0.5 * np.linalg.norm(
-                np.cross(
-                    np.array(facet_coords[0]) - np.array(facet_coords[2]),
-                    np.array(facet_coords[1]) - np.array(facet_coords[2]),
-                )
-            )
-
-            area_interstice = 1 - packed_area / triangle_area  # in fraction
-            area_interstice_list.append(area_interstice if area_interstice > 0 else 0)
-        return area_interstice_list
-
-    @staticmethod
-    def analyze_vol_interstice(center_coords, nn_coords, center_r, nn_rs, convex_hull_simplices):
-        """Analyze the volume interstices in the tetrahedra formed by center
-        atom and neighbor convex hull triplets.
-        Args:
-            center_coords ([float]): Central atomic coordinates.
-            nn_coords (array-like, shape (N, 3)): Nearest Neighbors' coordinates
-            center_r (float): central atom's radius.
-            nn_rs ([float]): Nearest Neighbors' radii.
-            convex_hull_simplices (array-like, shape (M, 3)): Indices of points
-                forming the simplicial facets of convex hull.
-        Returns:
-            volume_interstice_list ([float]): Volume interstice list.
-        """
-        volume_interstice_list = list()
-
-        triplet_set = [(0, 1, 2), (1, 0, 2), (2, 0, 1)]
-        for facet_indices in convex_hull_simplices:
-            facet_coords = nn_coords[facet_indices]
-            facet_rs = nn_rs[facet_indices]
-            solid_angles = list()
-            for triplet in triplet_set:
-                s_a = solid_angle(
-                    facet_coords[triplet[0]],
-                    np.array(
-                        [
-                            facet_coords[triplet[1]],
-                            facet_coords[triplet[2]],
-                            center_coords,
-                        ]
-                    ),
-                )
-                solid_angles.append(s_a)
-
-            # calculate neighbors' packed volume in the tetrahedron
-            packed_volume = 0
-            for s_a, nn_r in zip(solid_angles, facet_rs):
-                packed_volume += s_a / 3 * pow(nn_r, 3)
-
-            # add center atom's volume in the tetrahedron
-            center_solid_angle = solid_angle(center_coords, facet_coords)
-            packed_volume += center_solid_angle / 3 * pow(center_r, 3)
-
-            volume = vol_tetra(center_coords, *facet_coords)
-
-            volume_interstice = 1 - packed_volume / volume
-            volume_interstice_list.append(volume_interstice if volume_interstice > 0 else 0)
-        return volume_interstice_list
-
-    def feature_labels(self):
-        labels = list()
-        labels += ["Interstice_dist_%s" % stat for stat in self.stats] if "dist" in self.interstice_types else []
-        labels += ["Interstice_area_%s" % stat for stat in self.stats] if "area" in self.interstice_types else []
-        labels += ["Interstice_vol_%s" % stat for stat in self.stats] if "vol" in self.interstice_types else []
-        return labels
-
-    def citations(self):
-        return [
-            "@article{Wang2019,"
-            "title = {A transferable machine-learning framework linking "
-            "interstice distribution and plastic heterogeneity in metallic "
-            "glasses}, "
-            "author = {Qi Wang and Anubhav Jain},"
-            "journal = {Nature Communications}, year = {2019}, "
-            "pages = {5537}, volume = {10}, "
-            "doi = {10.1038/s41467-019-13511-9}, "
-            "url = {https://www.nature.com/articles/s41467-019-13511-9}}"
-        ]
-
-    def implementors(self):
-        return ["Qi Wang"]
-
-
-class ChemicalSRO(BaseFeaturizer):
-    """
-    Chemical short range ordering, deviation of local site and nominal structure compositions
-
-    Chemical SRO features to evaluate the deviation
-    of local chemistry with the nominal composition of the structure.
-
-    A local bonding preference is computed using
-    f_el = N_el/(sum of N_el) - c_el,
-    where N_el is the number of each element type in the neighbors around
-    the target site, sum of N_el is the sum of all possible element types
-    (coordination number), and c_el is the composition of the specific
-    element in the entire structure.
-    A positive f_el indicates the "bonding" with the specific element
-    is favored, at least in the target site;
-    A negative f_el indicates the "bonding" is not favored, at least
-    in the target site.
-
-    Note that ChemicalSRO is only featurized for elements identified by
-    "fit" (see following), thus "fit" must be called before "featurize",
-    or else an error will be raised.
-
-    Features:
-        CSRO__[nn method]_[element] - The Chemical SRO of a site computed based
-            on neighbors determined with a certain  NN-detection method for
-            a certain element.
-    """
-
-    def __init__(self, nn, includes=None, excludes=None, sort=True):
-        """Initialize the featurizer
-
-        Args:
-            nn (NearestNeighbor): instance of one of pymatgen's NearestNeighbor
-                                  classes.
-            includes (array-like or str): elements included to calculate CSRO.
-            excludes (array-like or str): elements excluded to calculate CSRO.
-            sort (bool): whether to sort elements by mendeleev number."""
-        self.nn = nn
-        self.includes = includes
-        if self.includes:
-            self.includes = [Element(el).symbol for el in np.atleast_1d(self.includes)]
-        self.excludes = excludes
-        if self.excludes:
-            self.excludes = [Element(el).symbol for el in np.atleast_1d(self.excludes)]
-        self.sort = sort
-        self.el_list_ = None
-        self.el_amt_dict_ = None
-
-    @staticmethod
-    def from_preset(preset, **kwargs):
-        """
-        Use one of the standard instances of a given NearNeighbor class.
-        Args:
-            preset (str): preset type ("VoronoiNN", "JmolNN",
-                          "MiniumDistanceNN", "MinimumOKeeffeNN",
-                          or "MinimumVIRENN").
-            **kwargs: allow to pass args to the NearNeighbor class.
-        Returns:
-            ChemicalSRO from a preset.
-        """
-        nn_ = getattr(pymatgen.analysis.local_env, preset)
-        return ChemicalSRO(nn_(**kwargs))
-
-    def fit(self, X, y=None):
-        """
-        Identify elements to be included in the following featurization,
-        by intersecting the elements present in the passed structures with
-        those explicitly included (or excluded) in __init__. Only elements
-        in the self.el_list_ will be featurized.
-        Besides, compositions of the passed structures will also be "stored"
-        in a dict of self.el_amt_dict_, avoiding repeated calculation of
-        composition when featurizing multiple sites in the same structure.
-        Args:
-            X (array-like): containing Pymatgen structures and sites, supports
-                            multiple choices:
-                            -2D array-like object:
-                             e.g. [[struct, site], [struct, site], …]
-                                  np.array([[struct, site], [struct, site], …])
-                            -Pandas dataframe:
-                             e.g. df[['struct', 'site']]
-            y : unused (added for consistency with overridden method signature)
-        Returns:
-            self
-        """
-        structs = np.atleast_2d(X)[:, 0]
-        if not all([isinstance(struct, Structure) for struct in structs]):
-            raise TypeError("This fit requires an array-like input of Pymatgen " "Structures and sites!")
-
-        self.el_amt_dict_ = {}
-        el_set_ = set()
-        for s in structs:
-            if str(s) not in self.el_amt_dict_.keys():
-                el_amt_ = s.composition.fractional_composition.get_el_amt_dict()
-                els_ = (
-                    set(el_amt_.keys())
-                    if self.includes is None
-                    else set([el for el in el_amt_.keys() if el in self.includes])
-                )
-                els_ = els_ if self.excludes is None else els_ - set(self.excludes)
-                if els_:
-                    self.el_amt_dict_[str(s)] = el_amt_
-                el_set_ = el_set_ | els_
-        self.el_list_ = sorted(list(el_set_), key=lambda el: Element(el).mendeleev_no) if self.sort else list(el_set_)
-        return self
-
-    def featurize(self, struct, idx):
-        """
-        Get CSRO features of site with given index in input structure.
-        Args:
-            struct (Structure): Pymatgen Structure object.
-            idx (int): index of target site in structure.
-        Returns:
-            (list of floats): Chemical SRO features for each element.
-        """
-
-        check_is_fitted(self, ["el_amt_dict_", "el_list_"])
-
-        csro = [0.0] * len(self.el_list_)
-        if str(struct) in self.el_amt_dict_.keys():
-            el_amt = self.el_amt_dict_[str(struct)]
-            nn_el_amt = dict.fromkeys(el_amt, 0)
-            nn_list = self.nn.get_nn(struct, idx)
-            for nn in nn_list:
-                if str(nn.specie.symbol) in self.el_list_:
-                    nn_el_amt[str(nn.specie.symbol)] += 1 / len(nn_list)
-            for el in el_amt.keys():
-                if el in self.el_list_:
-                    csro[self.el_list_.index(el)] = nn_el_amt[el] - el_amt[el]
-        return csro
-
-    def feature_labels(self):
-        check_is_fitted(self, ["el_amt_dict_", "el_list_"])
-
-        return ["CSRO_{}_{}".format(el, self.nn.__class__.__name__) for el in self.el_list_]
-
-    def citations(self):
-        citations = []
-        if self.nn.__class__.__name__ == "VoronoiNN":
-            citations.append(
-                "@article{voronoi_jreineangewmath_1908, title={"
-                "Nouvelles applications des param\\`{e}tres continus \\`{a} la "
-                "th'{e}orie des formes quadratiques. Sur quelques "
-                "propri'{e}t'{e}s des formes quadratiques positives"
-                ' parfaites}, journal={Journal f"ur die reine und angewandte '
-                "Mathematik}, number={133}, pages={97-178}, year={1908}}"
-            )
-            citations.append(
-                "@article{dirichlet_jreineangewmath_1850, title={"
-                '"{U}ber die Reduction der positiven quadratischen Formen '
-                "mit drei unbestimmten ganzen Zahlen}, journal={Journal "
-                'f"ur die reine und angewandte Mathematik}, number={40}, '
-                "pages={209-227}, doi={10.1515/crll.1850.40.209}, year={1850}}"
-            )
-        if self.nn.__class__.__name__ == "JmolNN":
-            citations.append(
-                "@misc{jmol, title = {Jmol: an open-source Java "
-                "viewer for chemical structures in 3D}, howpublished = {"
-                "\\url{http://www.jmol.org/}}}"
-            )
-        if self.nn.__class__.__name__ == "MinimumOKeeffeNN":
-            citations.append(
-                "@article{okeeffe_jamchemsoc_1991, title={Atom "
-                "sizes and bond lengths in molecules and crystals}, journal="
-                "{Journal of the American Chemical Society}, author={"
-                "O'Keeffe, M. and Brese, N. E.}, number={113}, pages={"
-                "3226-3229}, doi={doi:10.1021/ja00009a002}, year={1991}}"
-            )
-        if self.nn.__class__.__name__ == "MinimumVIRENN":
-            citations.append(
-                "@article{shannon_actacryst_1976, title={"
-                "Revised effective ionic radii and systematic studies of "
-                "interatomic distances in halides and chalcogenides}, "
-                "journal={Acta Crystallographica}, author={Shannon, R. D.}, "
-                "number={A32}, pages={751-767}, doi={"
-                "10.1107/S0567739476001551}, year={1976}"
-            )
-        if self.nn.__class__.__name__ in [
-            "MinimumDistanceNN",
-            "MinimumOKeeffeNN",
-            "MinimumVIRENN",
-        ]:
-            citations.append(
-                "@article{zimmermann_frontmater_2017, "
-                "title={Assessing local structure motifs using order "
-                "parameters for motif recognition, interstitial "
-                "identification, and diffusion path characterization}, "
-                "journal={Frontiers in Materials}, author={Zimmermann, "
-                "N. E. R. and Horton, M. K. and Jain, A. and Haranczyk, M.}, "
-                "number={4:34}, doi={10.3389/fmats.2017.00034}, year={2017}}"
-            )
-        return citations
-
-    def implementors(self):
-        return ["Qi Wang"]
-
-
-class GaussianSymmFunc(BaseFeaturizer):
-    """
-    Gaussian symmetry function features suggested by Behler et al.
-
-    The function is based on pair distances and angles, to approximate the functional
-    dependence of local energies, originally used in the fitting of
-    machine-learning potentials.
-    The symmetry functions can be divided to a set of radial functions
-    (g2 function), and a set of angular functions (g4 function).
-    The number of symmetry functions returned are based on parameters
-    of etas_g2, etas_g4, zetas_g4 and gammas_g4.
-    See the original papers for more details:
-    “Atom-centered symmetry functions for constructing high-dimensional
-    neural network potentials”, J Behler, J Chem Phys 134, 074106 (2011).
-    The cutoff function is taken as the polynomial form (cosine_cutoff)
-    to give a smoothed truncation.
-    A Fortran and a different Python version can be found in the code
-    Amp: Atomistic Machine-learning Package
-    (https://bitbucket.org/andrewpeterson/amp).
-    Args:
-        etas_g2 (list of floats): etas used in radial functions.
-                                  (default: [0.05, 4., 20., 80.])
-        etas_g4 (list of floats): etas used in angular functions.
-                                  (default: [0.005])
-        zetas_g4 (list of floats): zetas used in angular functions.
-                                   (default: [1., 4.])
-        gammas_g4 (list of floats): gammas used in angular functions.
-                                    (default: [+1., -1.])
-        cutoff (float): cutoff distance. (default: 6.5)
-    """
-
-    def __init__(self, etas_g2=None, etas_g4=None, zetas_g4=None, gammas_g4=None, cutoff=6.5):
-        self.etas_g2 = etas_g2 if etas_g2 else [0.05, 4.0, 20.0, 80.0]
-        self.etas_g4 = etas_g4 if etas_g4 else [0.005]
-        self.zetas_g4 = zetas_g4 if zetas_g4 else [1.0, 4.0]
-        self.gammas_g4 = gammas_g4 if gammas_g4 else [+1.0, -1.0]
-        self.cutoff = cutoff
-
-    @staticmethod
-    def cosine_cutoff(rs, cutoff):
-        """
-        Polynomial cutoff function to give a smoothed truncation of the Gaussian
-        symmetry functions.
-        Args:
-            rs (ndarray): distances to elements
-            cutoff (float): cutoff distance.
-        Returns:
-            (ndarray) cutoff function.
-        """
-        cutoff_fun = 0.5 * (np.cos(np.pi * rs / cutoff) + 1.0)
-        cutoff_fun[rs > cutoff] = 0
-        return cutoff_fun
-
-    @staticmethod
-    def g2(eta, rs, cutoff):
-        """
-        Gaussian radial symmetry function of the center atom,
-        given an eta parameter.
-        Args:
-            eta: radial function parameter.
-            rs: distances from the central atom to each neighbor
-            cutoff (float): cutoff distance.
-        Returns:
-            (float) Gaussian radial symmetry function.
-        """
-        ridge = np.exp(-eta * (rs ** 2.0) / (cutoff ** 2.0)) * GaussianSymmFunc.cosine_cutoff(rs, cutoff)
-        return ridge.sum()
-
-    @staticmethod
-    def g4(etas, zetas, gammas, neigh_dist, neigh_coords, cutoff):
-        """
-        Gaussian angular symmetry function of the center atom,
-        given a set of eta, zeta and gamma parameters.
-        Args:
-            eta ([float]): angular function parameters.
-            zeta ([float]): angular function parameters.
-            gamma ([float]): angular function parameters.
-            neigh_coords (list of [floats]): coordinates of neighboring atoms, with respect
-                to the central atom
-            cutoff (float): cutoff parameter.
-        Returns:
-            (float) Gaussian angular symmetry function for all combinations of eta, zeta, gamma
-        """
-
-        output = np.zeros((len(etas) * len(zetas) * len(gammas),))
-
-        # Loop over each neighbor j
-        for j, neigh_j in enumerate(neigh_coords):
-
-            # Compute the distance of each neighbor (k) to r
-            r_ij = neigh_dist[j]
-            d_jk = neigh_coords[(j + 1) :] - neigh_coords[j]
-            r_jk = np.linalg.norm(d_jk, 2, axis=1)
-            r_ik = neigh_dist[(j + 1) :]
-
-            # Compute the cosine term
-            cos_theta = np.dot(neigh_coords[(j + 1) :], neigh_coords[j]) / r_ij / r_ik
-
-            # Compute the cutoff function (independent of eta/zeta/gamma)
-            cutoff_fun = (
-                GaussianSymmFunc.cosine_cutoff(np.array([r_ij]), cutoff)
-                * GaussianSymmFunc.cosine_cutoff(r_ik, cutoff)
-                * GaussianSymmFunc.cosine_cutoff(r_jk, cutoff)
-            )
-
-            # Compute the g4 for each combination of eta/gamma/zeta
-            ind = 0
-            for eta in etas:
-                # Compute the eta term
-                eta_term = np.exp(-eta * (r_ij ** 2.0 + r_ik ** 2.0 + r_jk ** 2.0) / (cutoff ** 2.0)) * cutoff_fun
-                for zeta in zetas:
-                    for gamma in gammas:
-                        term = (1.0 + gamma * cos_theta) ** zeta * eta_term
-                        output[ind] += term.sum() * 2.0 ** (1.0 - zeta)
-                        ind += 1
-        return output
-
-    def featurize(self, struct, idx):
-        """
-        Get Gaussian symmetry function features of site with given index
-        in input structure.
-        Args:
-            struct (Structure): Pymatgen Structure object.
-            idx (int): index of target site in structure.
-        Returns:
-            (list of floats): Gaussian symmetry function features.
-        """
-        gaussian_funcs = []
-
-        # Get the neighbors within the cutoff
-        neighbors = struct.get_neighbors(struct[idx], self.cutoff)
-
-        # Get coordinates of the neighbors, relative to the central atom
-        neigh_coords = np.subtract([neigh[0].coords for neigh in neighbors], struct[idx].coords)
-
-        # Get the distances for later use
-        neigh_dists = np.array([neigh[1] for neigh in neighbors])
-
-        # Compute all G2
-        for eta_g2 in self.etas_g2:
-            gaussian_funcs.append(self.g2(eta_g2, neigh_dists, self.cutoff))
-
-        # Compute all G4s
-        gaussian_funcs.extend(
-            GaussianSymmFunc.g4(
-                self.etas_g4,
-                self.zetas_g4,
-                self.gammas_g4,
-                neigh_dists,
-                neigh_coords,
-                self.cutoff,
-            )
-        )
-        return gaussian_funcs
-
-    def feature_labels(self):
-        return ["G2_{}".format(eta_g2) for eta_g2 in self.etas_g2] + [
-            "G4_{}_{}_{}".format(eta_g4, zeta_g4, gamma_g4)
-            for eta_g4 in self.etas_g4
-            for zeta_g4 in self.zetas_g4
-            for gamma_g4 in self.gammas_g4
-        ]
-
-    def citations(self):
-        gsf_citation = (
-            "@Article{Behler2011, author = {Jörg Behler}, "
-            "title = {Atom-centered symmetry functions for constructing "
-            "high-dimensional neural network potentials}, "
-            "journal = {The Journal of Chemical Physics}, year = {2011}, "
-            "volume = {134}, number = {7}, pages = {074106}, "
-            "doi = {10.1063/1.3553717}}"
-        )
-        amp_citation = (
-            "@Article{Khorshidi2016, "
-            "author = {Alireza Khorshidi and Andrew A. Peterson}, "
-            "title = {Amp : A modular approach to machine learning in "
-            "atomistic simulations}, "
-            "journal = {Computer Physics Communications}, year = {2016}, "
-            "volume = {207}, pages = {310--324}, "
-            "doi = {10.1016/j.cpc.2016.05.010}}"
-        )
-        return [gsf_citation, amp_citation]
-
-    def implementors(self):
-        return ["Qi Wang"]
-
-
-class EwaldSiteEnergy(BaseFeaturizer):
-    """
-    Compute site energy from Coulombic interactions
-
-    User notes:
-        - This class uses that `charges that are already-defined for the structure`.
-        - Ewald summations can be expensive. If you evaluating every site in many
-          large structures, run all of the sites for each structure at the same time.
-          We cache the Ewald result for the structure that was run last, so looping
-          over sites and then structures is faster than structures than sites.
-    Features:
-        ewald_site_energy - Energy for the site computed from Coulombic interactions"""
-
-    def __init__(self, accuracy=None):
-        """
-        Args:
-            accuracy (int): Accuracy of Ewald summation, number of decimal places
-        """
-        self.accuracy = accuracy
-
-        # Variables used then caching the Ewald result
-        self.__last_structure = None
-        self.__last_ewald = None
-
-    def featurize(self, strc, idx):
-        """
-        Args:
-            struct (Structure): Pymatgen Structure object.
-            idx (int): index of target site in structure.
-        Returns:
-            ([float]) - Electrostatic energy of the site
-        """
-
-        # Check if the new input is the last
-        #  Note: We use 'is' rather than structure comparisons for speed
-        if strc is self.__last_structure:
-            ewald = self.__last_ewald
-        else:
-            self.__last_structure = strc
-            ewald = EwaldSummation(strc, acc_factor=self.accuracy)
-            self.__last_ewald = ewald
-        return [ewald.get_site_energy(idx)]
-
-    def feature_labels(self):
-        return ["ewald_site_energy"]
-
-    def implementors(self):
-        return ["Logan Ward"]
-
-    def citations(self):
-        return [
-            "@Article{Ewald1921,"
-            "author = {Ewald, P. P.},"
-            "doi = {10.1002/andp.19213690304},"
-            "issn = {00033804},"
-            "journal = {Annalen der Physik},"
-            "number = {3},"
-            "pages = {253--287},"
-            "title = {{Die Berechnung optischer und elektrostatischer Gitterpotentiale}},"
-            "url = {http://doi.wiley.com/10.1002/andp.19213690304},"
-            "volume = {369},"
-            "year = {1921}"
-            "}"
-        ]
-
-
-class CoordinationNumber(BaseFeaturizer):
-    """
-    Number of first nearest neighbors of a site.
-
-    Determines the number of nearest neighbors of a site using one of
-    pymatgen's NearNeighbor classes. These nearest neighbor calculators
-    can return weights related to the proximity of each neighbor to this
-    site. It is possible to take these weights into account to prevent
-    the coordination number from changing discontinuously with small
-    perturbations of a structure, either by summing the total weights
-    or using the normalization method presented by
-    [Ward et al.](http://link.aps.org/doi/10.1103/PhysRevB.96.014107)
-
-    Features:
-        CN_[method] - Coordination number computed using a certain method
-            for calculating nearest neighbors.
-    """
-
-    @staticmethod
-    def from_preset(preset, **kwargs):
-        """
-        Use one of the standard instances of a given NearNeighbor class.
-        Args:
-            preset (str): preset type ("VoronoiNN", "JmolNN",
-                          "MiniumDistanceNN", "MinimumOKeeffeNN",
-                          or "MinimumVIRENN").
-            **kwargs: allow to pass args to the NearNeighbor class.
-        Returns:
-            CoordinationNumber from a preset.
-        """
-        nn_ = getattr(pymatgen.analysis.local_env, preset)
-        return CoordinationNumber(nn_(**kwargs))
-
-    def __init__(self, nn=None, use_weights="none"):
-        """Initialize the featurizer
-
-        Args:
-            nn (NearestNeighbor) - Method used to determine coordination number
-            use_weights (string) - Method used to account for weights of neighbors:
-                'none' - Do not use weights when computing coordination number
-                'sum' - Use sum of weights as the coordination number
-                'effective' - Compute the 'effective coordination number', which
-                    is computed as :math:`\\frac{(\sum_n w_n)^2)}{\sum_n w_n^2}`
-        """
-        self.nn = nn or VoronoiNN()
-        self.use_weights = use_weights
-
-    def featurize(self, struct, idx):
-        """
-        Get coordintion number of site with given index in input
-        structure.
-        Args:
-            struct (Structure): Pymatgen Structure object.
-            idx (int): index of target site in structure struct.
-        Returns:
-            [float] - Coordination number
-        """
-        if self.use_weights is None or self.use_weights == "none":
-            return [self.nn.get_cn(struct, idx, use_weights=False)]
-        elif self.use_weights == "sum":
-            return [self.nn.get_cn(struct, idx, use_weights=True)]
-        elif self.use_weights == "effective":
-            # TODO: Should this weighting code go in pymatgen? I'm not sure if it even necessary to distinguish it from the 'sum' method -lw
-            nns = get_nearest_neighbors(self.nn, struct, idx)
-            weights = [n["weight"] for n in nns]
-            return [np.sum(weights) ** 2 / np.sum(np.power(weights, 2))]
-        else:
-            raise ValueError("Weighting method not recognized: " + str(self.use_weights))
-
-    def feature_labels(self):
-        # TODO: Should names contain weighting scheme? -lw
-        return ["CN_{}".format(self.nn.__class__.__name__)]
-
-    def citations(self):
-        citations = []
-        if self.nn.__class__.__name__ == "VoronoiNN":
-            citations.append(
-                "@article{voronoi_jreineangewmath_1908, title={"
-                "Nouvelles applications des param\\`{e}tres continus \\`{a} la "
-                "th'{e}orie des formes quadratiques. Sur quelques "
-                "propri'{e}t'{e}s des formes quadratiques positives"
-                ' parfaites}, journal={Journal f"ur die reine und angewandte '
-                "Mathematik}, number={133}, pages={97-178}, year={1908}}"
-            )
-            citations.append(
-                "@article{dirichlet_jreineangewmath_1850, title={"
-                '"{U}ber die Reduction der positiven quadratischen Formen '
-                "mit drei unbestimmten ganzen Zahlen}, journal={Journal "
-                'f"ur die reine und angewandte Mathematik}, number={40}, '
-                "pages={209-227}, doi={10.1515/crll.1850.40.209}, year={1850}}"
-            )
-        if self.nn.__class__.__name__ == "JmolNN":
-            citations.append(
-                "@misc{jmol, title = {Jmol: an open-source Java "
-                "viewer for chemical structures in 3D}, howpublished = {"
-                "\\url{http://www.jmol.org/}}}"
-            )
-        if self.nn.__class__.__name__ == "MinimumOKeeffeNN":
-            citations.append(
-                "@article{okeeffe_jamchemsoc_1991, title={Atom "
-                "sizes and bond lengths in molecules and crystals}, journal="
-                "{Journal of the American Chemical Society}, author={"
-                "O'Keeffe, M. and Brese, N. E.}, number={113}, pages={"
-                "3226-3229}, doi={doi:10.1021/ja00009a002}, year={1991}}"
-            )
-        if self.nn.__class__.__name__ == "MinimumVIRENN":
-            citations.append(
-                "@article{shannon_actacryst_1976, title={"
-                "Revised effective ionic radii and systematic studies of "
-                "interatomic distances in halides and chalcogenides}, "
-                "journal={Acta Crystallographica}, author={Shannon, R. D.}, "
-                "number={A32}, pages={751-767}, doi={"
-                "10.1107/S0567739476001551}, year={1976}"
-            )
-        if self.nn.__class__.__name__ in [
-            "MinimumDistanceNN",
-            "MinimumOKeeffeNN",
-            "MinimumVIRENN",
-        ]:
-            citations.append(
-                "@article{zimmermann_frontmater_2017, "
-                "title={Assessing local structure motifs using order "
-                "parameters for motif recognition, interstitial "
-                "identification, and diffusion path characterization}, "
-                "journal={Frontiers in Materials}, author={Zimmermann, "
-                "N. E. R. and Horton, M. K. and Jain, A. and Haranczyk, M.}, "
-                "number={4:34}, doi={10.3389/fmats.2017.00034}, year={2017}}"
-            )
-        return citations
-
-    def implementors(self):
-        return ["Nils E. R. Zimmermann", "Logan Ward"]
-
-
-class GeneralizedRadialDistributionFunction(BaseFeaturizer):
-    """
-    Compute the general radial distribution function (GRDF) for a site.
-
-    The GRDF is a radial measure of crystal order around a site. There are two
-    featurizing modes:
-
-    1. GRDF: (recommended) - n_bins length vector
-        In GRDF mode, The GRDF is computed by considering all sites around a
-        central site (i.e., no sites are omitted when computing the GRDF). The
-        features output from this mode will be vectors with length n_bins.
-
-    2. pairwise GRDF: (advanced users) - n_bins x n_sites matrix
-        In this mode, GRDFs are are still computed around a central site, but
-        only one other site (and their translational equivalents) are used to
-        compute a GRDF (e.g. site 1 with site 2 and the translational
-        equivalents of site 2). This results in a a n_sites x n_bins matrix of
-        features. Requires `fit` for determining the max number of sites for
-
-    The GRDF is a generalization of the partial radial distribution function
-    (PRDF). In contrast with the PRDF, the bins of the GRDF are not mutually-
-    exclusive and need not carry a constant weight of 1. The PRDF is a case of
-    the GRDF when the bins are rectangular functions. Examples of other
-    functions to use with the GRDF are Gaussian, trig, and Bessel functions.
-
-    See :func:`~matminer.featurizers.utils.grdf` for a full list of available binning functions.
-
-    There are two preset conditions:
-        gaussian: bin functions are gaussians
-        histogram: bin functions are rectangular functions
-
-    Args:
-        bins:   ([AbstractPairwise]) List of pairwise binning functions. Each of these functions
-            must implement the AbstractPairwise class.
-        cutoff: (float) maximum distance to look for neighbors
-        mode:   (str) the featurizing mode. supported options are:
-                    'GRDF' and 'pairwise_GRDF'
-    """
-
-    def __init__(self, bins, cutoff=20.0, mode="GRDF"):
-        self.bins = bins
-        self.cutoff = cutoff
-
-        if mode not in ["GRDF", "pairwise_GRDF"]:
-            raise AttributeError("{} is not a valid GRDF mode. try " '"GRDF" or "pairwise_GRDF"'.format(mode))
-        else:
-            self.mode = mode
-
-        self.fit_labels = None
-
-    def fit(self, X, y=None, **fit_kwargs):
-        """
-        Determine the maximum number of sites in X to assign correct feature
-        labels
-
-        Args:
-            X - [list of tuples], training data
-                tuple values should be (struc, idx)
-        Returns:
-            self
-        """
-
-        max_sites = max([len(X[i][0]._sites) for i in range(len(X))])
-        self.fit_labels = ["site2 {} {}".format(i, bin.name()) for bin in self.bins for i in range(max_sites)]
-        return self
-
-    def featurize(self, struct, idx):
-        """
-        Get GRDF of the input structure.
-        Args:
-            struct (Structure): Pymatgen Structure object.
-            idx (int): index of target site in structure struct.
-
-        Returns:
-            Flattened list of GRDF values. For each run mode the list order is:
-                GRDF:          bin#
-                pairwise GRDF: site2# bin#
-            The site2# corresponds to a pymatgen site index and bin#
-            corresponds to one of the bin functions
-        """
-
-        if not struct.is_ordered:
-            raise ValueError("Disordered structure support not built yet")
-
-        # Get list of neighbors by site
-        # Indexing is [site#][neighbor#][pymatgen Site, distance, site index]
-        sites = struct._sites
-        central_site = sites[idx]
-        neighbors_lst = struct.get_neighbors(central_site, self.cutoff, include_index=True)
-        sites = range(0, len(sites))
-
-        # Generate lists of pairwise distances according to run mode
-        if self.mode == "GRDF":
-            # Make a single distance collection
-            distance_collection = [[neighbor[1] for neighbor in neighbors_lst]]
-        else:
-            # Make pairwise distance collections for pairwise GRDF
-            distance_collection = [
-                [neighbor[1] for neighbor in neighbors_lst if neighbor[2] == site_idx] for site_idx in sites
-            ]
-
-        # compute bin counts for each list of pairwise distances
-        bin_counts = []
-        for values in distance_collection:
-            bin_counts.append([sum(bin(values)) for bin in self.bins])
-
-        # Compute "volume" of each bin to normalize GRDFs
-        volumes = [bin.volume(self.cutoff) for bin in self.bins]
-
-        # normalize the bin counts by the bin volume to compute features
-        features = []
-        for values in bin_counts:
-            features.extend(np.array(values) / np.array(volumes))
-
-        return features
-
-    def feature_labels(self):
-        if self.mode == "GRDF":
-            return [bin.name() for bin in self.bins]
-        else:
-            if self.fit_labels:
-                return self.fit_labels
-            else:
-                raise AttributeError("the fit method must be called first, to " "determine the correct feature labels.")
-
-    @staticmethod
-    def from_preset(preset, width=1.0, spacing=1.0, cutoff=10, mode="GRDF"):
-        """
-        Preset bin functions for this featurizer. Example use:
-            >>> GRDF = GeneralizedRadialDistributionFunction.from_preset('gaussian')
-            >>> GRDF.featurize(struct, idx)
-
-        Args:
-            preset (str): shape of bin (either 'gaussian' or 'histogram')
-            width (float): bin width. std dev for gaussian, width for histogram
-            spacing (float): the spacing between bin centers
-            cutoff (float): maximum distance to look for neighbors
-            mode (str): featurizing mode. either 'GRDF' or 'pairwise_GRDF'
-        """
-
-        # Generate bin functions
-        if preset == "gaussian":
-            bins = []
-            for center in np.arange(0.0, cutoff, spacing):
-                bins.append(Gaussian(width, center))
-        elif preset == "histogram":
-            bins = []
-            for start in np.arange(0, cutoff, spacing):
-                bins.append(Histogram(start, width))
-        else:
-            raise ValueError("Not a valid preset condition.")
-        return GeneralizedRadialDistributionFunction(bins, cutoff=cutoff, mode=mode)
-
-    def citations(self):
-        return [
-            "@article{PhysRevB.95.144110, title = {Representation of compo"
-            "unds for machine-learning prediction of physical properties},"
-            " author = {Seko, Atsuto and Hayashi, Hiroyuki and Nakayama, "
-            "Keita and Takahashi, Akira and Tanaka, Isao},"
-            "journal = {Phys. Rev. B}, volume = {95}, issue = {14}, "
-            "pages = {144110}, year = {2017}, publisher = {American Physic"
-            "al Society}, doi = {10.1103/PhysRevB.95.144110}}"
-        ]
-
-    def implementors(self):
-        return ["Maxwell Dylla", "Saurabh Bajaj", "Logan Williams"]
-
-
-class AngularFourierSeries(BaseFeaturizer):
-    """
-    Compute the angular Fourier series (AFS), including both angular and radial info
-
-    The AFS is the product of pairwise distance function (g_n, g_n') between two pairs
-    of atoms (sharing the common central site) and the cosine of the angle
-    between the two pairs. The AFS is a 2-dimensional feature (the axes are g_n,
-    g_n').
-
-    Examples of distance functionals are square functions, Gaussian, trig
-    functions, and Bessel functions. An example for Gaussian:
-        lambda d: exp( -(d - d_n)**2 ), where d_n is the coefficient for g_n
-
-    See :func:`~matminer.featurizers.utils.grdf` for a full list of available binning functions.
-
-    There are two preset conditions:
-        gaussian: bin functions are gaussians
-        histogram: bin functions are rectangular functions
-
-    Features:
-        AFS ([gn], [gn']) - Angular Fourier Series between binning functions (g1 and g2)
-
-    Args:
-        bins:   ([AbstractPairwise]) a list of binning functions that
-                implement the AbstractPairwise base class
-        cutoff: (float) maximum distance to look for neighbors. The
-                 featurizer will run slowly for large distance cutoffs
-                 because of the number of neighbor pairs scales as
-                 the square of the number of neighbors
-    """
-
-    def __init__(self, bins, cutoff=10.0):
-        self.bins = bins
-        self.cutoff = cutoff
-
-    def featurize(self, struct, idx):
-        """
-        Get AFS of the input structure.
-        Args:
-            struct (Structure): Pymatgen Structure object.
-            idx (int): index of target site in structure struct.
-
-        Returns:
-            Flattened list of AFS values. the list order is:
-                g_n g_n'
-        """
-
-        if not struct.is_ordered:
-            raise ValueError("Disordered structure support not built yet")
-
-        # Generate list of neighbor position vectors (relative to central
-        # atom) and distances from each central site as tuples
-        sites = struct._sites
-        central_site = sites[idx]
-        neighbors_lst = struct.get_neighbors(central_site, self.cutoff)
-        neighbor_collection = [(neighbor[0].coords - central_site.coords, neighbor[1]) for neighbor in neighbors_lst]
-
-        # Generate exhaustive permutations of neighbor pairs around each
-        # central site (order matters). Does not allow repeat elements (i.e.
-        # there are two distinct sites in every permutation)
-        neighbor_tuples = itertools.permutations(neighbor_collection, 2)
-
-        # Generate cos(theta) between neighbor pairs for each central site.
-        # Also, retain data on neighbor distances for each pair
-        # process with matrix algebra, we really need the speed here
-        data = np.array(list(neighbor_tuples))
-        v1, v2 = np.vstack(data[:, 0, 0]), np.vstack(data[:, 1, 0])
-        distances = data[:, :, 1]
-        neighbor_pairs = np.concatenate(
-            [
-                np.clip(
-                    np.einsum("ij,ij->i", v1, v2) / np.linalg.norm(v1, axis=1) / np.linalg.norm(v2, axis=1),
-                    -1.0,
-                    1.0,
-                ).reshape(-1, 1),
-                distances,
-            ],
-            axis=1,
-        )
-
-        # Generate distance functional matrix (g_n, g_n')
-        bin_combos = list(itertools.product(self.bins, repeat=2))
-
-        # Compute AFS values for each element of the bin matrix
-        # need to cast arrays as floats to use np.exp
-        cos_angles, dist1, dist2 = (
-            neighbor_pairs[:, 0].astype(float),
-            neighbor_pairs[:, 1].astype(float),
-            neighbor_pairs[:, 2].astype(float),
-        )
-        features = [sum(combo[0](dist1) * combo[1](dist2) * cos_angles) for combo in bin_combos]
-
-        return features
-
-    def feature_labels(self):
-        bin_combos = list(itertools.product(self.bins, repeat=2))
-        return ["AFS ({}, {})".format(combo[0].name(), combo[1].name()) for combo in bin_combos]
-
-    @staticmethod
-    def from_preset(preset, width=0.5, spacing=0.5, cutoff=10):
-        """
-        Preset bin functions for this featurizer. Example use:
-            >>> AFS = AngularFourierSeries.from_preset('gaussian')
-            >>> AFS.featurize(struct, idx)
-
-        Args:
-            preset (str): shape of bin (either 'gaussian' or 'histogram')
-            width (float): bin width. std dev for gaussian, width for histogram
-            spacing (float): the spacing between bin centers
-            cutoff (float): maximum distance to look for neighbors
-        """
-
-        # Generate bin functions
-        if preset == "gaussian":
-            bins = []
-            for center in np.arange(0.0, cutoff, spacing):
-                bins.append(Gaussian(width, center))
-        elif preset == "histogram":
-            bins = []
-            for start in np.arange(0, cutoff, spacing):
-                bins.append(Histogram(start, width))
-        else:
-            raise ValueError("Not a valid preset condition.")
-        return AngularFourierSeries(bins, cutoff=cutoff)
-
-    def citations(self):
-        return [
-            "@article{PhysRevB.95.144110, title = {Representation of compo"
-            "unds for machine-learning prediction of physical properties},"
-            " author = {Seko, Atsuto and Hayashi, Hiroyuki and Nakayama, "
-            "Keita and Takahashi, Akira and Tanaka, Isao},"
-            "journal = {Phys. Rev. B}, volume = {95}, issue = {14}, "
-            "pages = {144110}, year = {2017}, publisher = {American Physic"
-            "al Society}, doi = {10.1103/PhysRevB.95.144110}}"
-        ]
-
-    def implementors(self):
-        return ["Maxwell Dylla", "Logan Williams"]
-
-
-# TODO: Figure out whether to take NN-counting method as an option (see VoronoiFingerprint)
-class LocalPropertyDifference(BaseFeaturizer):
-    """
-    Differences in elemental properties between site and its neighboring sites.
-
-    Uses the Voronoi tessellation of the structure to determine the
-    neighbors of the site, and assigns each neighbor (:math:`n`) a
-    weight (:math:`A_n`) that corresponds to the area of the facet
-    on the tessellation corresponding to that neighbor.
-    The local property difference is then computed by
-    :math:`\\frac{\sum_n {A_n |p_n - p_0|}}{\sum_n {A_n}}`
-    where :math:`p_n` is the property (e.g., atomic number) of a neighbor
-    and :math:`p_0` is the property of a site. If signed parameter is assigned
-    True, signed difference of the properties is returned instead of absolute
-    difference.
-
-    Features:
-        - "local property difference in [property]" - Weighted average
-            of differences between an elemental property of a site and
-            that of each of its neighbors, weighted by size of face on
-            Voronoi tessellation
-
-    References:
-         `Ward et al. _PRB_ 2017 <http://link.aps.org/doi/10.1103/PhysRevB.96.024104>`_
-    """
-
-    def __init__(
-        self,
-        data_source=MagpieData(),
-        weight="area",
-        properties=("Electronegativity",),
-        signed=False,
-    ):
-        """Initialize the featurizer
-
-        Args:
-            data_source (AbstractData) - Class from which to retrieve
-                elemental properties
-            weight (str) - What aspect of each voronoi facet to use to
-                weigh each neighbor (see VoronoiNN)
-            properties ([str]) - List of properties to use (default=['Electronegativity'])
-            signed (bool) - whether to return absolute difference or signed difference of
-                            properties(default=False (absolute difference))
-        """
-        self.data_source = data_source
-        self.properties = properties
-        self.weight = weight
-        self.signed = signed
-
-    @staticmethod
-    def from_preset(preset):
-        """
-        Create a new LocalPropertyDifference class according to a preset
-
-        Args:
-            preset (str) - Name of preset
-        """
-
-        if preset == "ward-prb-2017":
-            return LocalPropertyDifference(
-                data_source=MagpieData(),
-                properties=[
-                    "Number",
-                    "MendeleevNumber",
-                    "AtomicWeight",
-                    "MeltingT",
-                    "Column",
-                    "Row",
-                    "CovalentRadius",
-                    "Electronegativity",
-                    "NsValence",
-                    "NpValence",
-                    "NdValence",
-                    "NfValence",
-                    "NValence",
-                    "NsUnfilled",
-                    "NpUnfilled",
-                    "NdUnfilled",
-                    "NfUnfilled",
-                    "NUnfilled",
-                    "GSvolume_pa",
-                    "GSbandgap",
-                    "GSmagmom",
-                    "SpaceGroupNumber",
-                ],
-            )
-        else:
-            raise ValueError("Unrecognized preset: " + preset)
-
-    def featurize(self, strc, idx):
-        # Get the targeted site
-        my_site = strc[idx]
-
-        # Get the tessellation of a site
-        nn = get_nearest_neighbors(VoronoiNN(weight=self.weight), strc, idx)
-
-        # Get the element and weight of each site
-        elems = [n["site"].specie for n in nn]
-        weights = [n["weight"] for n in nn]
-
-        # Compute the difference for each property
-        output = np.zeros((len(self.properties),))
-        total_weight = np.sum(weights)
-        for i, p in enumerate(self.properties):
-            my_prop = self.data_source.get_elemental_property(my_site.specie, p)
-            n_props = self.data_source.get_elemental_properties(elems, p)
-            if self.signed == False:
-                output[i] = np.dot(weights, np.abs(np.subtract(n_props, my_prop))) / total_weight
-            else:
-                output[i] = np.dot(weights, np.subtract(n_props, my_prop)) / total_weight
-
-        return output
-
-    def feature_labels(self):
-        if self.signed == False:
-            return ["local difference in " + p for p in self.properties]
-        else:
-            return ["local signed difference in " + p for p in self.properties]
-
-    def citations(self):
-        return [
-            "@article{Ward2017,"
-            "author = {Ward, Logan and Liu, Ruoqian "
-            "and Krishna, Amar and Hegde, Vinay I. "
-            "and Agrawal, Ankit and Choudhary, Alok "
-            "and Wolverton, Chris},"
-            "doi = {10.1103/PhysRevB.96.024104},"
-            "journal = {Physical Review B},"
-            "pages = {024104},"
-            "title = {{Including crystal structure attributes "
-            "in machine learning models of formation energies "
-            "via Voronoi tessellations}},"
-            "url = {http://link.aps.org/doi/10.1103/PhysRevB.96.014107},"
-            "volume = {96},year = {2017}}",
-            "@article{jong_chen_notestine_persson_ceder_jain_asta_gamst_2016,"
-            "title={A Statistical Learning Framework for Materials Science: "
-            "Application to Elastic Moduli of k-nary Inorganic Polycrystalline Compounds}, "
-            "volume={6}, DOI={10.1038/srep34256}, number={1}, journal={Scientific Reports}, "
-            "author={Jong, Maarten De and Chen, Wei and Notestine, Randy and Persson, "
-            "Kristin and Ceder, Gerbrand and Jain, Anubhav and Asta, Mark and Gamst, Anthony}, "
-            "year={2016}, month={Mar}}",
-        ]
-
-    def implementors(self):
-        return ["Logan Ward", "Aik Rui Tan"]
-
-
-class SiteElementalProperty(BaseFeaturizer):
-    """
-    Elemental properties of atom on a certain site
-
-    Features:
-        site [property] - Elemental property for this site
-
-    References:
-        `Seko et al., _PRB_ (2017) <http://link.aps.org/doi/10.1103/PhysRevB.95.144110>`_
-        `Schmidt et al., _Chem Mater_. (2017) <http://dx.doi.org/10.1021/acs.chemmater.7b00156>`_
-    """
-
-    def __init__(self, data_source=None, properties=("Number",)):
-        """Initialize the featurizer
-
-        Args:
-            data_source (AbstractData): Tool used to look up elemental properties
-            properties ([string]): List of properties to use for features
-        """
-        self.data_source = data_source or MagpieData()
-        self.properties = properties
-        self._preset_citations = []
-
-    def featurize(self, strc, idx):
-        # Get the site
-        site = strc[idx]
-
-        # Get the properties
-        elem = site.specie if isinstance(site.specie, Element) else site.specie.element
-        props = [self.data_source.get_elemental_property(elem, p) for p in self.properties]
-
-        return props
-
-    def feature_labels(self):
-        return ["site {}".format(p) for p in self.properties]
-
-    def citations(self):
-        return self._preset_citations
-
-    def implementors(self):
-        return ["Logan Ward"]
-
-    @staticmethod
-    def from_preset(preset):
-        """Create the class with pre-defined settings
-
-        Args:
-            preset (string): Desired preset
-        Returns:
-            SiteElementalProperty initialized with desired settings
-        """
-
-        if preset == "seko-prb-2017":
-            output = SiteElementalProperty(
-                data_source=MagpieData(),
-                properties=[
-                    "Number",
-                    "AtomicWeight",
-                    "Row",
-                    "Column",
-                    "FirstIonizationEnergy",
-                    "SecondIonizationEnergy",
-                    "ElectronAffinity",
-                    "Electronegativity",
-                    "AllenElectronegativity",
-                    "VdWRadius",
-                    "CovalentRadius",
-                    "AtomicRadius",
-                    "ZungerPP-r_s",
-                    "ZungerPP-r_p",
-                    "MeltingT",
-                    "BoilingT",
-                    "Density",
-                    "MolarVolume",
-                    "HeatFusion",
-                    "HeatVaporization",
-                    "LogThermalConductivity",
-                    "HeatCapacityMass",
-                ],
-            )
-            output._preset_citations.append(
-                "@article{Seko2017,"
-                "author = {Seko, Atsuto and Hayashi, Hiroyuki and "
-                "Nakayama, Keita and Takahashi, Akira and Tanaka, Isao},"
-                "doi = {10.1103/PhysRevB.95.144110},"
-                "journal = {Physical Review B}, number = {14},"
-                "pages = {144110},"
-                "title = {{Representation of compounds for machine-learning prediction of physical properties}},"
-                "url = {http://link.aps.org/doi/10.1103/PhysRevB.95.144110},"
-                "volume = {95}, year = {2017}}"
-            )
-            return output
-        else:
-            raise ValueError("Unrecognized preset: {}".format(preset))
 
 
 @lru_cache(maxsize=32)
